requirements.txt
pygame~=2.5.0
numpy~=1.23.5
neat-python
